Add asylum requests component
