+++ src/pages/asylumRequests/index.tsx
@@ -1,0 +1,3 @@
+export const AsylumRequests = () => {
+    return <p>Testing</p>
+}
