+++ src/routes/AppRoutes.tsx
@@ -1,4 +1,5 @@
 import { MainLayout } from "@/components/layout"
+import { AsylumRequests } from "@/pages/asylumRequests"
 import { Dashboard } from "@/pages/dashboard"
 import { BrowserRouter, Route, Routes } from "react-router-dom"
 
@@ -8,6 +9,7 @@ export const AppRoutes = () =>{
         <Routes>
             <Route element={<MainLayout/>}>
             <Route path="/" element={<Dashboard/>}/>
+            <Route path="/requests" element={<AsylumRequests/>}/>
             </Route>
         </Routes>
         </BrowserRouter>
